requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai==0.3.2
pillow==10.1.0
python-multipart==0.0.6 
